refactor(llamabot)🔧: Refactor message handling and API functions
- Remove unused imports and simplify message component imports.
- Centralize high-level API functions for message handling into the messages module.
- Ensure directory creation for .llamabot in the main init module.
